+++ src/components/button-toggle/__spec__.js
@@ -13,7 +13,29 @@ describe('ButtonToggle', () => {
 
   describe('mainClasses', () => {
     it('returns the classes for the component', () => {
-      expect(instance.mainClasses).toEqual('carbon-button-toggle common-input');
+      expect(instance.mainClasses).toEqual('carbon-button-toggle carbon-button-toggle--large common-input');
+    });
+
+    describe('when it is grouped', () => {
+      it('returns the grouped class', () => {
+        instance = TestUtils.renderIntoDocument(
+          <ButtonToggle size={ 'small' }><span>Plus</span></ButtonToggle>
+        );
+        expect(instance.mainClasses).toEqual(
+          'carbon-button-toggle carbon-button-toggle--small common-input'
+        );
+      });
+    });
+
+    describe('when the size is set to small', () => {
+      it('returns the small class', () => {
+        instance = TestUtils.renderIntoDocument(
+          <ButtonToggle grouped={ true }><span>Plus</span></ButtonToggle>
+        );
+        expect(instance.mainClasses).toEqual(
+          'carbon-button-toggle carbon-button-toggle--large carbon-button-toggle--grouped common-input'
+        );
+      });
     });
   });